Added Additional Product Support, now saving customer/employee in cookies and bug fixes

+++ app/models/Order.scala
@@ -5,5 +5,5 @@ package models
   * @author Kamil Gorszczyk
   */
 
-case class Order(var id: Long, customerID: Int, produktID: Int, amount: Int, extras: String, price: Double, orderTime: String, size: Double)
+case class Order(var id: Long, customerID: Long, produktID: Long, amount: Int, extras: String, price: Double, orderTime: String, size: Double)
